Restructure App.js, remove old account page
