--- client/src/components/Account/Account.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const Account = () => {
-    return (
-        <div style={{paddingTop:"50px"}} >
-            ACCOUNT
-        </div>
-    )
-}
-
-export default Account
