+++ client/src/components/App.js
@@ -5,25 +5,23 @@ import {connect} from 'react-redux'
 
 import { BrowserRouter, Switch, Route, Link, Redirect } from 'react-router-dom';
 
-import * as ROUTES from '../../routes';
+import * as ROUTES from '../routes';
 
-import Landing from '../Landing/Landing';
-import Account from '../Account/Account';
-import Logout from '../Logout/Logout';
-import NotFound from '../NotFound/NotFound';
-import Home from '../Home/Home';
+import Landing from './Landing/Landing';
+import Logout from './Logout/Logout';
+import NotFound from './NotFound/NotFound';
+import Home from './Home/Home';
 
-import ProtectedRoute from '../ProtectedRoute/ProtectedRoute';
-import UnprotectedRoute from '../UnprotectedRoute/UnprotectedRoute';
-import setUser from '../../actions/user'
+import ProtectedRoute from './ProtectedRoute/ProtectedRoute';
+import UnprotectedRoute from './UnprotectedRoute/UnprotectedRoute';
+import setUser from '../actions/user'
 import {bindActionCreators} from 'redux'
 
-import spinner from '../../images/spinner.gif'
-import SearchPage from '../Search/SearchPage';
-import UserPage from '../UserPage/User'
+import spinner from '../images/spinner.gif'
+import SearchPage from './Search/SearchPage';
+import UserPage from './UserPage/User'
 
 export function fetchIsAuth(setUser) {
-  // return dispatch => {
       fetch(ROUTES.API_IS_AUTH, {
         method: "GET",
         credentials: "same-origin"
@@ -36,7 +34,6 @@ export function fetchIsAuth(setUser) {
           setUser(json.authenticated, true, json.user.id, json.user.email, json.user.username, json.user.profilePicture, json.user.regDate)
         }
       })
-  // }
 }
 
 class App extends React.Component {
@@ -49,20 +46,6 @@ class App extends React.Component {
   componentDidMount() {
     const {setUser} = this.props.actions
     fetchIsAuth(setUser)
-    // const {setUser} = this.props.actions
-
-    // fetch(ROUTES.API_IS_AUTH, {
-    //   method: "GET",
-    //   credentials: "same-origin"
-    // })
-    // .then(res => res.json())
-    // .then(json => {
-    //   if (json.error) {
-    //     setUser(false, true)
-    //   } else {
-    //     setUser(json.authenticated, true, json.user.id, json.user.email, json.user.username, json.user.profilePicture, json.user.regDate)
-    //   }
-    // })
   }
 
   componentWillUnmount() {
@@ -85,9 +68,6 @@ class App extends React.Component {
               <Route exact path={ROUTES.HOME} >
                 <ProtectedRoute component={Home} />
               </Route>
-              {/* <Route exact path={ROUTES.ACCOUNT}>
-                <ProtectedRoute component={UserPage} />
-              </Route> */}
               <Route exact path={ROUTES.LOGOUT}>
                 <ProtectedRoute component={Logout} />
               </Route>
@@ -103,9 +83,6 @@ class App extends React.Component {
             null}
           </div>
         </BrowserRouter>
-
-        // :
-        // null
     )
   }
 }
